fix(tui): split effective_home_dir into a leaf so skill_cli can resolve it
tests/skill_cli.rs pulls network_policy.rs and skills/install.rs into the
test binary via #[path] includes, so `crate::` there is the test binary's
root rather than the lib. #4757 replaced dirs::home_dir() with
crate::config::effective_home_dir() in both files, leaving two E0433s and
a broken test build on every platform:

    error[E0433]: cannot find `config` in `crate`
      --> crates/tui/tests/../src/network_policy.rs:324:27
      --> crates/tui/tests/../src/skills/install.rs:64:12

The obvious fix — include config/paths.rs behind a `mod config` shim —
does not work. Integration test binaries compile with cfg(test) set, so
paths.rs brings its `crate::test_support` call with it, and test_support
brings `crate::config_persistence` behind that. Shimming each layer would
put three stand-ins for production code in the test, including a fake
env mutex, which is a worse test than none.

effective_home_dir was already a leaf over std/dirs; it just shared a file
with the cfg(test)-gated env_config_path. Move it to config/home.rs, which
has no crate:: references at all, and re-export from paths.rs so
config::effective_home_dir and every `use paths::{...}` caller resolve
unchanged. The test binary includes that one file and gets the real
implementation rather than a divergent shim.

skill_cli is the only affected test binary; the other #[path]-including
tests pull files with no crate::config references.
